requirements.txt
amqp==2.2.2
billiard==3.5.0.3
celery==4.0.2
certifi==2017.11.5
chardet==3.0.4
Django==2.0.1
django-cors-headers==2.1.0
django-datatables-view==1.14.0
djangorestframework==3.7.7
idna==2.6
kombu==4.1.0
numpy==1.14.0
pandas==0.22.0
psycopg2==2.7.3.2
python-bittrex==0.2.2
python-dateutil==2.6.1
pytz==2017.3
redis==2.10.6
requests==2.18.4
selenium==3.8.0
six==1.11.0
splinter==0.7.7
TA-Lib==0.4.14
urllib3==1.22
vine==1.1.4
